auth scheme ok

+++ api/views.py
@@ -1,6 +1,11 @@
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from django.conf import settings
+
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication, TokenAuthentication
+
 from .serializers import TaskSerializer
 from .models import Task
 
@@ -31,6 +36,7 @@ def taskList(request):
 
 
 @api_view(['GET'])
+
 def taskDetail(request,pk):
     """
     API endpoint for detailed view on a specific task
@@ -41,18 +47,20 @@ def taskDetail(request,pk):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def taskCreate(request):
     """
     API endpoint for task creation
     """
     serializer = TaskSerializer(data=request.data)
     if serializer.is_valid():
-        serializer.save()
+        serializer.save(user=request.user)
         return Response(serializer.data, status=201)
-    return Response(serializer.erros, status=400)
+    return Response({'error. serializer not valid'}, status=400)
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def taskUpdate(request,pk):
     """
     API endpoint for task update
@@ -64,11 +72,18 @@ def taskUpdate(request,pk):
     return Response(serializer.data)
 
 
-@api_view(['POST'])
+@api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def taskDelete(request,pk):
     """
     API endpoint for task delete
     """
-    task = Task.objects.get(id=pk)
+    qs = Task.objects.filter(id=pk)
+    if not qs.exists():
+        return Response({}, status=400)
+    qs = qs.filter(user=request.user)
+    if not qs.exists():
+        return Response({'message':'Sorry, you are NOT allowed to delete this item'}, status=400)
+    task = qs.first()
     task.delete()
     return Response('Item Removed')
